implement client handler for start radio command
